+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - The [emmet-vim](https://github.com/mattn/emmet-vim) plugin is replaced by [cmp-emmet](https://github.com/jackieaskins/cmp-emmet).
+- Detection of `prettier` config files has been improved.
+- The `eslint` and `prettier` save actions are removed.
+- A new `PrettierFormatAll` command is created.
 
 ## [1.16.0] - 2022-12-20